refactor: Address PR feedback on GuardrailsConfiguration
- Add name field to GuardrailsConfiguration record
- Update getName() to include config type (e.g., "Guardrails (Guyton-Klinger)")
- Move GuardrailsConfiguration from domain.calculator to domain.config
- Update all imports and tests accordingly
