+++ .idea/codeStyles/Project.xml
@@ -1,10 +1,7 @@
 <component name="ProjectCodeStyleConfiguration">
   <code_scheme name="Project" version="173">
-    <!-- General settings -->
     <option name="RIGHT_MARGIN" value="120" />
     <option name="FORMATTER_TAGS_ENABLED" value="true" />
-
-    <!-- Java-specific settings -->
     <JavaCodeStyleSettings>
       <option name="GENERATE_FINAL_LOCALS" value="true" />
       <option name="GENERATE_FINAL_PARAMETERS" value="true" />
@@ -29,69 +26,19 @@
         </value>
       </option>
     </JavaCodeStyleSettings>
-
-    <!-- Indentation and formatting -->
     <codeStyleSettings language="JAVA">
       <option name="RIGHT_MARGIN" value="120" />
-      <option name="KEEP_LINE_BREAKS" value="true" />
       <option name="KEEP_FIRST_COLUMN_COMMENT" value="false" />
       <option name="KEEP_CONTROL_STATEMENT_IN_ONE_LINE" value="false" />
       <option name="KEEP_BLANK_LINES_IN_DECLARATIONS" value="1" />
       <option name="KEEP_BLANK_LINES_IN_CODE" value="1" />
       <option name="KEEP_BLANK_LINES_BEFORE_RBRACE" value="0" />
-      <option name="BLANK_LINES_AFTER_CLASS_HEADER" value="0" />
-      <option name="BLANK_LINES_AFTER_ANONYMOUS_CLASS_HEADER" value="0" />
-
-      <!-- K&R Brace Style -->
-      <option name="BRACE_STYLE" value="1" />
-      <option name="CLASS_BRACE_STYLE" value="1" />
-      <option name="METHOD_BRACE_STYLE" value="1" />
-      <option name="LAMBDA_BRACE_STYLE" value="1" />
-
-      <!-- Always use braces -->
       <option name="IF_BRACE_FORCE" value="3" />
       <option name="DOWHILE_BRACE_FORCE" value="3" />
       <option name="WHILE_BRACE_FORCE" value="3" />
       <option name="FOR_BRACE_FORCE" value="3" />
-
-      <!-- Spaces -->
-      <option name="SPACE_BEFORE_METHOD_CALL_PARENTHESES" value="false" />
-      <option name="SPACE_BEFORE_METHOD_PARENTHESES" value="false" />
-      <option name="SPACE_BEFORE_IF_PARENTHESES" value="true" />
-      <option name="SPACE_BEFORE_WHILE_PARENTHESES" value="true" />
-      <option name="SPACE_BEFORE_FOR_PARENTHESES" value="true" />
-      <option name="SPACE_BEFORE_CATCH_PARENTHESES" value="true" />
-      <option name="SPACE_BEFORE_SWITCH_PARENTHESES" value="true" />
-      <option name="SPACE_BEFORE_SYNCHRONIZED_PARENTHESES" value="true" />
-      <option name="SPACE_BEFORE_CLASS_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_METHOD_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_IF_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_ELSE_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_WHILE_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_FOR_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_DO_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_SWITCH_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_TRY_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_CATCH_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_FINALLY_LBRACE" value="true" />
-      <option name="SPACE_BEFORE_SYNCHRONIZED_LBRACE" value="true" />
-
-      <!-- Indentation -->
       <indentOptions>
-        <option name="INDENT_SIZE" value="4" />
         <option name="CONTINUATION_INDENT_SIZE" value="4" />
-        <option name="TAB_SIZE" value="4" />
-        <option name="USE_TAB_CHARACTER" value="false" />
-        <option name="SMART_TABS" value="false" />
-      </indentOptions>
-    </codeStyleSettings>
-
-    <!-- XML settings -->
-    <codeStyleSettings language="XML">
-      <indentOptions>
-        <option name="INDENT_SIZE" value="4" />
-        <option name="TAB_SIZE" value="4" />
-        <option name="USE_TAB_CHARACTER" value="false" />
       </indentOptions>
     </codeStyleSettings>
   </code_scheme>
